Add embed mode flag
This commit updates TIPSY to become aware when it is running inside
embedded mode and provided a pathbuilder to inspect. This is done by
adding an appropriate variable to state, and setting it automatically.

This embedded mode now provides a notice to the user that the
pathbuilder being inspected has been preloaded. It also prevents the
user from closing the pathbuilder, or loading another one.